Rename pipeline_storage file
