+++ graphrag/index/run/utils.py
@@ -33,7 +33,7 @@ from graphrag.index.config.storage import (
 from graphrag.index.context import PipelineRunContext, PipelineRunStats
 from graphrag.index.input import load_input
 from graphrag.index.storage.memory_pipeline_storage import MemoryPipelineStorage
-from graphrag.index.storage.typing import PipelineStorage
+from graphrag.index.storage.pipeline_storage import PipelineStorage
 from graphrag.logging import ProgressReporter
 
 log = logging.getLogger(__name__)
