+++ graphrag/index/run/workflow.py
@@ -19,7 +19,7 @@ from graphrag.callbacks.progress_workflow_callbacks import ProgressWorkflowCallb
 from graphrag.index.context import PipelineRunContext
 from graphrag.index.emit.table_emitter import TableEmitter
 from graphrag.index.run.profiling import _write_workflow_stats
-from graphrag.index.storage.typing import PipelineStorage
+from graphrag.index.storage.pipeline_storage import PipelineStorage
 from graphrag.index.typing import PipelineRunResult
 from graphrag.logging import ProgressReporter
 from graphrag.utils.storage import _load_table_from_storage
