+++ graphrag/index/storage/__init__.py
@@ -7,7 +7,7 @@ from .blob_pipeline_storage import BlobPipelineStorage, create_blob_storage
 from .file_pipeline_storage import FilePipelineStorage
 from .load_storage import load_storage
 from .memory_pipeline_storage import MemoryPipelineStorage
-from .typing import PipelineStorage
+from .pipeline_storage import PipelineStorage
 
 __all__ = [
     "BlobPipelineStorage",
